requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1
python-docx==1.1.2
google-cloud-aiplatform==1.66.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
